use useScripts and useExternalStyles hooks to dynamically load asciinema script
