+++ website/src/hooks/useExternalStyles.ts
@@ -1,6 +1,7 @@
 import { useEffect } from "react";
 
 export const useExternalStyles = (href: string,) => {
+
     useEffect(() => {
         const link = document.createElement('link');
         link.type = "text/css";
