+++ website/src/hooks/useScript.ts
@@ -1,16 +1,58 @@
-import { useEffect } from "react";
+import { useEffect, useState } from 'react'
 
-export const useScript = (url: string, isAsync?: boolean) => {
-    useEffect(() => {
-        const script = document.createElement('script');
-        script.type = "text/javascript";
-        script.src = url;
-        script.async = isAsync || false;
-        document.body.appendChild(script);
+export type Status = 'idle' | 'loading' | 'ready' | 'error'
+export type ScriptElement = HTMLScriptElement | null
 
-        return () => {
-            document.body.removeChild(script);
+export const useScript = (src: string): Status => {
+
+  const [status, setStatus] = useState<Status>(src ? 'loading' : 'idle')
+
+  useEffect(
+    () => {
+      if (!src) {
+        setStatus('idle')
+        return
+      }
+
+      let script: ScriptElement = document.querySelector(`script[src="${src}"]`)
+
+      if (!script) {
+        script = document.createElement('script')
+        script.src = src
+        script.async = true
+        script.setAttribute('data-status', 'loading')
+        document.body.appendChild(script)
+
+        const setAttributeFromEvent = (event: Event) => {
+          script?.setAttribute(
+            'data-status',
+            event.type === 'load' ? 'ready' : 'error',
+          )
         }
-    }, [url, isAsync]);
+
+        script.addEventListener('load', setAttributeFromEvent)
+        script.addEventListener('error', setAttributeFromEvent)
+      } else {
+        setStatus(script.getAttribute('data-status') as Status)
+      }
+
+      const setStateFromEvent = (event: Event) => {
+        setStatus(event.type === 'load' ? 'ready' : 'error')
+      }
+
+      script.addEventListener('load', setStateFromEvent)
+      script.addEventListener('error', setStateFromEvent)
+
+      return () => {
+        if (script) {
+          script.removeEventListener('load', setStateFromEvent)
+          script.removeEventListener('error', setStateFromEvent)
+        }
+      }
+    },
+    [src],
+  )
+
+  return status
 }
